test(npm): move resolver core tests to test_resolver.py
Resolves https://github.com/hermetoproject/hermeto/issues/1156

Move resolver-core npm tests from the mixed test_npm.py module to
tests/unit/package_managers/npm/test_resolver.py.

Relocate _resolve_npm validation/flow tests and lockfile/package.json
rewrite helper tests so resolver behavior is tested in one module.

No behavior changes; only test relocation and import adjustments.
